+++ tests/js/browser_contract.test.js
@@ -95,6 +95,8 @@ const MODULES = [
       "formatNumber",
       "formatKpi",
       "reportSummaryTotals",
+      "reportSecondaryStats",
+      "reportStatLabel",
     ],
     moved: [
       "reportsPeriodLabel",
@@ -113,6 +115,11 @@ const MODULES = [
       "flagParamLabel",
       "formatFlagParam",
       "reportSummaryTotals",
+      "reportSecondaryStats",
+      "reportStatLabel",
+      "reportTotalsBlock",
+      "collectReportStats",
+      "collectReportStatChildren",
     ],
   },
   {